Add yFiles dependency and render basic graphComponent canvas in the ProjectView
The yFiles package is added in development mode to the package.json and package-lock.json and the yFiles bundle/license files are specifically added to the .gitignore file, since they must not be published. For further development, the ProjectView component is set up to display a graphComponent canvas, where the nodess will be rendered.